変数名などの修正
