+++ Pin.rb
@@ -1,24 +1,5 @@
-#最大スコアを設定
-class MaxPin
+class Pin
   
-  def initialize(score,frame, throw_number)
-    @score = score
-    @frame = frame
-    @throw_number = throw_number
-  end
-  
-  MAX_PIN = 10
-  def build_max_pins(score, frame, throw_number)
-    # 最終フレーム以外(ストライク時は10-10 = 0となるので問題なし)
-    return MAX_PIN - score.sum unless @frame == 10
-    # 最終フレーム(1-2投目 ※下記はすべて最終フレームが入る)
-    return MAX_PIN - score.sum if throw_number < 3
-    # 最終フレーム(3投目)
-    return MAX_PIN if throw_number == 3
-  end
-end
-
-class Pins
   def initialize(frame, score, max_throw_number)
     # インスタンス変数には@をつける
     @frame = frame
@@ -28,6 +9,7 @@ class Pins
   
   def build_get_pins(frame, score, max_throw_number)
     
+    require './MaxPin.rb'
     max_throw_number.each do |throw_number| 
       
       max_pin = MaxPin.new(score, frame, throw_number)
